+++ finish/gateway/pom.xml
@@ -130,7 +130,7 @@
                         <configuration>
                             <jvmOptions>
                                 <param>-DGATEWAY_JOB_BASE_URI=http://localhost:9082</param>
-                                <param>-Dio.openliberty.guides.gateway.client.InventoryGatewayClient/mp-rest/url=http://localhost:9082</param>
+                                <param>-Dio.openliberty.guides.gateway.client.InventoryClient/mp-rest/url=http://localhost:9082</param>
                             </jvmOptions>
                         </configuration>
                     </execution>
