+++ finish/gateway/src/main/java/io/openliberty/guides/gateway/GatewayInventoryResource.java
@@ -10,32 +10,44 @@
  *     IBM Corporation - Initial implementation
  *******************************************************************************/
 // end::copyright[]
-package io.openliberty.guides.gateway.client;
+package io.openliberty.guides.gateway;
 
 import java.util.Properties;
 
+import javax.enterprise.context.RequestScoped;
+import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 
-import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+import org.eclipse.microprofile.rest.client.inject.RestClient;
 
+import io.openliberty.guides.gateway.client.InventoryClient;
 import io.openliberty.guides.models.InventoryList;
+import io.openliberty.guides.models.SystemData;
 
-@RegisterRestClient(baseUri = "http://inventory-service:9080")
-@Path("/inventory")
-public interface InventoryGatewayClient {
+@RequestScoped
+@Path("/systems")
+public class GatewayInventoryResource {
+
+    @Inject
+    @RestClient
+    private InventoryClient inventoryClient;
 
     @GET
-    @Path("systems")
     @Produces(MediaType.APPLICATION_JSON)
-    public InventoryList getInventory();
+    public InventoryList getSystems() {
+        return inventoryClient.getInventory();
+    }
 
     @GET
-    @Path("systems/{hostname}")
+    @Path("{hostname}")
     @Produces(MediaType.APPLICATION_JSON)
-    public Properties getProperties(@PathParam("hostname") String hostname);
+    public SystemData getSystem(@PathParam("hostname") String hostname) {
+        Properties properties = inventoryClient.getProperties(hostname);
+        return new SystemData(hostname, properties);
+    }
 
 }
